+++ src/components/NavBar/NavBar.tsx
@@ -10,6 +10,9 @@ import chekIsItEdit from './SubComponentFromNavBar/EditOrTextField';
 import checkIsItLogin from './SubComponentFromNavBar/LoginOrEmpty';
 import checkIsLogin from './SubComponentFromNavBar/AuthenOrSave';
 import loadable from '@loadable/component';
+import createGraph from '../../services/api/CreateGraphService';
+import isGraphExist from '../../services/api/CheckGraphExistsService';
+import updateGraph from '../../services/api/UpdateGraphService';
 
 /* eslint-disable @typescript-eslint/promise-function-async */
 const HomeIconButton = loadable(() => import('./NavBarButton/HomeIconButton'));
@@ -20,27 +23,6 @@ const UserIcon = loadable(() => import('./NavBarButton/UserIconButton'));
 export interface NavbarProps {
 	currentPage: string;
 	forceLogin?: boolean;
-}
-
-interface Graph {
-	id: number;
-	name: string;
-	preview: string;
-	owner: string;
-	created: string;
-	updated: string;
-}
-
-interface GraphDetailResponse {
-	status: number;
-	message: string;
-	data: Graph;
-}
-
-interface GraphValidationErrorResponse {
-	status: number;
-	message: string;
-	data: Graph;
 }
 
 export default function Navbar(props: NavbarProps): React.ReactElement {
@@ -73,86 +55,19 @@ export default function Navbar(props: NavbarProps): React.ReactElement {
 		setIsLogin(!isLogIn);
 	};
 
-	const isGraphExist = async (graphId: number): Promise<boolean> => {
-		try {
-			const response = await fetch(`http://127.0.0.1:8000/api/viewset/graphs/${graphId}/`, {
-				method: 'GET',
-				mode: 'cors',
-				headers: {
-					'Content-Type': 'application/json',
-				},
-			});
-
-			if (response.status === 200) {
-				return true; // Graph exists
-			} else if (response.status === 404) {
-				return false; // Graph does not exist
-			}
-			console.log(response);
-		} catch (e) {
-			console.log(e);
-		}
-		return false;
-	};
-
-	const updateGraph = async (graphId: number): Promise<void> => {
-		const dateTimeNow = new Date();
-		const graphReq = {
-			name: buttonText,
-			preview: 'minio_bucket_test2',
-			updated: dateTimeNow.toDateString(),
-			owner: auth.currentUser?.uid,
-		};
-		const response = await fetch(`http://127.0.0.1:8000/api/viewset/graphs/${graphId}/`, {
-			method: 'PUT',
-			mode: 'cors',
-			headers: {
-				'Content-Type': 'application/json',
-			},
-			body: JSON.stringify(graphReq),
-		});
-		if (!response.ok) {
-			const errorResponse: GraphValidationErrorResponse = await response.json();
-			throw new Error(errorResponse.message);
-		}
-		const graphDetail: GraphDetailResponse = await response.json();
-		console.log(graphDetail.data);
-	};
-
-	const createGraph = async (): Promise<void> => {
-		const graphReq = {
-			name: buttonText,
-			preview: 'minio_bucket_test',
-			owner: auth.currentUser?.uid,
-		};
-		const response = await fetch('http://127.0.0.1:8000/api/viewset/graphs/', {
-			method: 'POST',
-			mode: 'cors',
-			headers: {
-				'Content-Type': 'application/json',
-			},
-			body: JSON.stringify(graphReq),
-		});
-		if (!response.ok) {
-			const errorResponse: GraphValidationErrorResponse = await response.json();
-			throw new Error(errorResponse.message);
-		}
-		const graphDetail: GraphDetailResponse = await response.json();
-		setGid(graphDetail.data.id);
-		console.log(graphDetail.data);
-	};
-
 	const handleSaveIconClick = async (): Promise<void> => {
 		const user = auth.currentUser;
 		setIsSave(true);
 		if (user != null) {
+			const uid = user.uid;
 			console.log(user);
 			const isExist = await isGraphExist(gid);
 			console.log(isExist);
 			if (isExist) {
-				await updateGraph(gid);
+				await updateGraph(buttonText, gid, uid);
 			} else {
-				await createGraph();
+				const newGraphId = await createGraph(buttonText, uid);
+				setGid(newGraphId);
 			}
 		} else {
 			console.log('null user');
